Update web view frame when super view changes its size.
For some reason the 'autoresizingMask' property does not work as expected
when a Cordova project is built with Xcode 7.3 (latest iOS 9 SDK) and deployed
on device running on iOS 10. This results in web view not updating its frame
on rotation events.

In order to resolve the issue, implement 'viewWillTransitionToSize:withTransitionCoordinator:'
(iOS 8.0 and later) and layout the web view with updated frame.